Reorganized crate, adde feature flags
I think I had a fundamental misunderstanding about the difference
between phenotype and genotype. I've removed basically alll of my old
code that was supposed to operate on non-standard ga/p primitives.
Instead, If you want to use any of the algorithms you need to implement
the trait `Individual` that requires `From<BitVec>` and `Into<BitVec>`
(in the case of GA) or `From<Graph>` and `Into<Graph>` for gp.

I've also stubbed the other primtivies module and algorithms module.
I've been reading how the python library DEAP does things and am
realizing the error of my ways. By committing to specific types of
primitives, It will be much easier to optimize the algorithms.