Complete restructure (again) back to many types
I've restructured the project again, hopefully for the last time. I'm
going with the evolutionary algorithm taking many(8) type parameters.
Additionally, I've also made it a struct so that the algorithm and state
can be serialized and dumped to a file. This will help if the engine
crashes during evolution