Working operators for Vec<u32>, many types version
This is the first branch with a working GA. The genetic operators are
working for Vec<u32> right now. Right now I'm implementing everything
individually, but I think that this is wasting rust's powerful type
system. I'll be writing BitFlip to work on any "chromosome" that
supports a bitwise negation

I also want to remove individual as a struct and instead just have
selection take in a vec of tuples (fitness, genome)

Finally, I want the eval function be able to be a closure