complete rewrite, read notes
I was not happy with the way the crate was organized and the API. I've
completely rewritten the crate and removed everything that was currently
unimplemented to clean up. I've also moved all operators to a single
file that also includes the traits that they implement. This means that
its no longer necessary to import the trait to use each operator